backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.1
python-jose==3.3.0
python-multipart==0.0.9
pydantic==2.9.1
slowapi==0.1.9
httpx==0.27.2
pytest==8.3.3
supabase==2.5.0